+++ src/redux/actions/favourite.js
@@ -5,8 +5,9 @@ import { favorite, URL } from "../action-types/index.js";
 export function getFavorite(idUser) {
   return (dispatch) => {
     axios.get(`${URL}/favorites/${idUser}`)
+    
       .then(response => {dispatch({
-        type: favorite.GET_FAVORITE,
+        type: favorite.GET_FAVOURITE,
         payload: response.data
       })})
       .catch(error => {
